Gbfs

+++ include/InputHandler.h
@@ -10,11 +10,11 @@ class InputHandler
 {
 public:
     static SearchAlgorithm selectAlgorithm(const std::string& input);
-    static std::vector<std::vector<int>> processInput(int argc, char* argv[]);
+    static std::vector<std::vector<uint8_t>> processInput(int argc, char* argv[]);
 
 private:
-    static std::vector<int> parseState(const std::string& state);
-    static std::vector<std::vector<int>> readFromStdin();
+    static std::vector<uint8_t> parseState(const std::string& state);
+    static std::vector<std::vector<uint8_t>> readFromStdin();
 };
 
 #endif
